Seperate the Create Competition servlet to send the competition ID back to front end after the competition is created.
